+++ src/main/java/Event.java
@@ -1,2 +1,12 @@
-package PACKAGE_NAME;public class Event {
+public class Event extends Task{
+    protected String place;
+
+    public Event(String d, String place) {
+        super(d);
+        this.place = place;
+    }
+    @Override
+    public String toString(){
+        return "[E]" + super.toString() + " (at: " + place + ") ";
+    }
 }
