--- src/main/java/Level3.java
@@ -1,2 +1,0 @@
-package PACKAGE_NAME;public class Level3 {
-}
