--- src/main/java/Level4/Event.java
@@ -1,12 +1,0 @@
-public class Event extends Task{
-    protected String place;
-
-    public Event(String d, String place) {
-        super(d);
-        this.place = place;
-    }
-    @Override
-    public String toString(){
-        return "[E]" + super.toString() + " (at: " + place + ") ";
-    }
-}
