--- src/main/java/Level5/Deadline.java
@@ -1,13 +1,0 @@
-public class Deadline extends Task {
-    protected String due;
-
-    public Deadline(String d, String due){
-        super(d);
-        this.due = due;
-    }
-    @Override
-    public String toString() {
-        return "[D]" + super.toString() + " (by: " + due +") ";
-    }
-
-}
