--- src/main/java/Level5/ToDos.java
@@ -1,13 +1,0 @@
-import java.util.Arrays;
-
-public class ToDos extends Task {
-
-    public ToDos(String d) {
-        super(d);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
