+++ src/main/java/ToDos.java
@@ -1,2 +1,13 @@
-package PACKAGE_NAME;public class ToDos {
+import java.util.Arrays;
+
+public class ToDos extends Task {
+
+    public ToDos(String d) {
+        super(d);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
 }
